Modify carbon cycle to calculate alpha term
The Mimi version of FAIR previously treated alpha (state-dependent carbon uptake scaling factor) as an exogenous parameter read in from data.  This was just to get the Mimi version of FAIR up and running and to make sure it matched the original Python equations/code.  The Mimi version now calculates the alpha term for each time period.